working on alphabetControl
