+++ VK/VK/Storage/GroupStorage.swift
@@ -1,0 +1,17 @@
+//
+//  GroupStorage.swift
+//  VK
+//
+//  Created by Lera on 26.06.22.
+//
+
+import Foundation
+import UIKit
+
+var groups = [
+    Group(name: "Animal Planet", image: UIImage(named: "animalPlanet")),
+    Group(name: "Discovery Knowledge", image: UIImage(named: "discoveryKnowledge")),
+    Group(name: "Discovery Science", image: UIImage(named: "discoveryScience")),
+    Group(name: "Investigation Discovery", image: UIImage(named: "investigationDiscovery")),
+    Group(name: "Military History", image: UIImage(named: "militaryHistory")),
+]
